refactoring ProjectsList.js
