+++ front-end/src/components/DeleteProject.js
@@ -3,59 +3,6 @@ import { Link } from 'react-router-dom';
 import axios from 'axios';
 import ProjectDataService from "../services/ProjectDataService";
 
-export default class DeleteProject extends Component {
-    constructor(props) {
-        super(props);
+export default function DeleteProject() {
 
-        this.deleteExercise = this.deleteExercise.bind(this)
-
-        this.state = {exercises: []};
-    }
-
-    componentDidMount() {
-        axios.get('http://localhost:5000/exercises/')
-            .then(response => {
-                this.setState({ exercises: response.data })
-            })
-            .catch((error) => {
-                console.log(error);
-            })
-    }
-
-    deleteExercise(id) {
-        axios.delete('http://localhost:5000/exercises/'+id)
-            .then(response => { console.log(response.data)});
-
-        this.setState({
-            exercises: this.state.exercises.filter(el => el._id !== id)
-        })
-    }
-
-    exerciseList() {
-        return this.state.exercises.map(currentexercise => {
-            return <Exercise exercise={currentexercise} deleteExercise={this.deleteExercise} key={currentexercise._id}/>;
-        })
-    }
-
-    render() {
-        return (
-            <div>
-                <h3>Logged Exercises</h3>
-                <table className="table">
-                    <thead className="thead-light">
-                    <tr>
-                        <th>Username</th>
-                        <th>Description</th>
-                        <th>Duration</th>
-                        <th>Date</th>
-                        <th>Actions</th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    { this.exerciseList() }
-                    </tbody>
-                </table>
-            </div>
-        )
-    }
 }
